Add stepper into transactions block
