+++ components/add-transactions/add-transactions-block.tsx
@@ -1,8 +1,7 @@
 "use client";
 
 import { useEffect, useState } from "react";
-import { FilePond } from "react-filepond";
-import type { FilePondFile, FilePondInitialFile } from "filepond";
+import type { FilePondInitialFile } from "filepond";
 import {
   DATES_CSV_FORMAT_OPTIONS,
   URL_UPLOAD_TRANSACTION_FILE,
@@ -13,6 +12,9 @@ import { AddTransactionsTable } from "../tables/add-transactions-tables/add-tran
 import { columns } from "../tables/add-transactions-tables/columns";
 import { ScrollArea } from "../ui/scroll-area";
 import { useAddTransactionTable } from "@/hooks/use-add-transaction-table";
+import { Separator } from "../ui/separator";
+import { Stepper } from "./stepper";
+import { InputFileBlock } from "./input-file-block";
 
 type AddTransactionsBlockProps = {
   userCategories: Categories[];
@@ -21,95 +23,31 @@ type AddTransactionsBlockProps = {
 export const AddTransactionsBlock = ({
   userCategories,
 }: AddTransactionsBlockProps) => {
+  const [currentStep, setCurrentStep] = useState(0);
   const [files, setFiles] = useState<Array<FilePondInitialFile | File | Blob>>(
     [],
   );
   const [CSVDateFormat, setCSVDateFormat] = useState<string>(
     DATES_CSV_FORMAT_OPTIONS[0],
   );
-  const [isReady, setIsReady] = useState(false);
-  const { toast } = useToast();
   const { addTransactions, setAddTransactions, setUserCategories } =
     useAddTransactionTable();
-
-  useEffect(() => {
-    setIsReady(true);
-  }, []);
 
   useEffect(() => {
     setUserCategories(userCategories);
   }, [userCategories]);
 
-  const handleUpdateFiles = (fileItems: FilePondFile[]) => {
-    const updatedFiles: Array<FilePondInitialFile | File | Blob> =
-      fileItems.map((fileItem) => fileItem.file);
-    setFiles(updatedFiles);
-  };
-
-  const handleFileProcessed = (response: any) => {
-    const { ok: responseOk, data }: ResponseFile = JSON.parse(response);
-    if (responseOk && data) {
-      setAddTransactions((prevState) => {
-        const parsedTrans = data.map((trans, i) => ({
-          ...trans,
-          id: prevState.length + i,
-        }));
-        return [...prevState, ...parsedTrans];
-      });
-      return "success";
-    }
-    return "failure";
-  };
-
   return (
     <>
-      <ScrollArea maxHeight={154} className="my-2">
-        <div
-          className="file-wrapper max-w-[400px] mx-auto -mb-4"
-          style={isReady ? { opacity: 1 } : undefined}
-        >
-          <FilePond
-            files={files}
-            allowMultiple
-            maxFiles={3}
-            maxParallelUploads={1}
-            onupdatefiles={handleUpdateFiles}
-            credits={false}
-            labelIdle='Drag & Drop your CSV files or <span class="filepond--label-action"> Browse </span>'
-            labelFileProcessing="Parsing"
-            labelFileProcessingComplete="Parse completed"
-            labelTapToCancel="Please wait..."
-            name="files"
-            server={{
-              process: {
-                url: URL_UPLOAD_TRANSACTION_FILE,
-                method: "POST",
-                withCredentials: false,
-                onload: handleFileProcessed,
-                onerror: (response) => {
-                  try {
-                    const parsedRes = JSON.parse(response);
-                    toast({
-                      title: "There was an error uploading the file",
-                      description: parsedRes.error,
-                      variant: "destructive",
-                    });
-                  } catch (error) {
-                    // Meaning its not an object stringified
-                    console.log("ERROR UPLOADING THE CSV:", response);
-                    toast({
-                      title: "There was an error uploading the file",
-                      description: `Please, contact to support or try again later`,
-                      variant: "destructive",
-                    });
-                  }
-                },
-              },
-            }}
-          />
-        </div>
-      </ScrollArea>
-      {addTransactions.length > 0 && (
+      {/* <ScrollArea maxHeight={154} className="my-2"> */}
+      <Stepper currentStep={currentStep} />
+      <Separator />
+      {currentStep === 0 && (
+        <InputFileBlock files={files} setFiles={setFiles} />
+      )}
+      {/* </ScrollArea> */}
+
+      {/* {addTransactions.length > 0 && (
         <div className="relative">
           <AddTransactionsTable
             columns={columns}
@@ -117,7 +55,7 @@ export const AddTransactionsBlock = ({
             dataLength={addTransactions.length}
           />
         </div>
-      )}
+      )} */}
     </>
   );
 };
